update export default
- added generation of voximplant-kit-sdk.d.ts
- transfer of types to a separate module
