+++ clear.js
@@ -1,0 +1,4 @@
+const fs = require('fs');
+
+
+fs.rmdirSync('dist', { recursive: true });
