+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "@voximplant/kit-functions-sdk",
-  "version": "0.0.33",
+  "version": "0.0.34",
   "description": "Voximplant Kit cloud functions client",
   "main": "dist/index.js",
   "typescript": {
@@ -8,7 +8,8 @@
   },
   "typings": "dist/index.d.ts",
   "scripts": {
-    "test": "jest --silent"
+    "test": "jest --silent",
+    "build": "node clear.js && tsc && node bundle.js"
   },
   "private": false,
   "repository": {
@@ -29,6 +30,7 @@
     "querystring": "^0.2.0"
   },
   "devDependencies": {
-    "jest": "^26.6.3"
+    "jest": "^26.6.3",
+    "dts-bundle": "^0.7.3"
   }
 }
